General: Fix compiler warnings

+++ Day1.h
@@ -18,7 +18,7 @@ int Part1(const std::vector<int>& data) {
 
 int Part2(const std::vector<int>& data) {
 	int sum{0};
-	for (int i{0}; i < data.size(); i++) {
+	for (size_t i{0}; i < data.size(); i++) {
 		if (i + 3 < data.size() && data[i + 3] > data[i]) {
 			++sum;
 		}
